+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+- Run commands inside the container using `exec`
+```
+aww exec sign -- echo 1 2 3
+```
+
 ## 0.1.0 - 2018-08-07
 
 ### Added